# docs/swagger.yaml
basePath: /
definitions:
  model.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
  req.ReqSignIn:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    type: object
  req.ReqSignUp:
    properties:
      email:
        type: string
      fullName:
        description: tags
        type: string
      password:
        type: string
    required:
    - email
    - fullName
    type: object
  req.ReqUpdateUser:
    properties:
      email:
        type: string
      fullName:
        description: tags
        type: string
    required:
    - email
    - fullName
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: More
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Github Trending API
  version: "1.0"
paths:
  /user/profile:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - jwt: []
      summary: get user profile
      tags:
      - user-service
  /user/profile/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.ReqUpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - jwt: []
      summary: get user profile
      tags:
      - user-service
  /user/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.ReqSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Sign in to access your account
      tags:
      - user-service
  /user/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.ReqSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Create new account
      tags:
      - user-service
securityDefinitions:
  jwt:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"